requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=4.5=1_gnu
_pytorch_select=0.2=gpu_0
antlr-python-runtime=4.8=pyhd8ed1ab_3
backcall=0.2.0=pyhd3eb1b0_0
blas=1.0=mkl
blosc=1.20.1=hd408876_0
boost=1.67.0=py37_4
brotli=1.0.9=he6710b0_2
brotlipy=0.7.0=py37h27cfd23_1003
brunsli=0.1=h2531618_0
bzip2=1.0.8=h7b6447c_0
ca-certificates=2021.10.26=h06a4308_2
caffe=1.0=py37hbab4207_5
cairo=1.14.12=h8948797_3
certifi=2021.10.8=py37h06a4308_2
cffi=1.14.4=py37h261ae71_0
chardet=4.0.0=py37h06a4308_1003
charls=2.1.0=he6710b0_2
click=7.1.2=pyhd3eb1b0_0
cloudpickle=1.6.0=py_0
cryptography=3.3.1=py37h3c74f83_0
cudatoolkit=11.0.221=h6bb024c_0
cudnn=6.0=0
cycler=0.10.0=py37_0
cytoolz=0.11.0=py37h7b6447c_0
dask-core=2021.1.1=pyhd3eb1b0_0
dbus=1.13.18=hb2f20db_0
decorator=4.4.2=pyhd3eb1b0_0
dominate=2.4.0=py_0
exifread=2.1.2=py_1
expat=2.2.10=he6710b0_2
ffmpeg=4.0=hcdf2ecd_0
fontconfig=2.13.1=h6c09931_0
freeglut=3.0.0=hf484d3e_5
freetype=2.10.4=h5ab3b9f_0
gflags=2.2.2=he6710b0_0
giflib=5.1.4=h14c3975_1
glib=2.66.1=h92f7085_0
glog=0.3.5=hf484d3e_1
graphite2=1.3.14=h23475e2_0
gst-plugins-base=1.14.0=h8213a91_2
gstreamer=1.14.0=h28cd5cc_2
h5py=2.8.0=py37h989c5e5_3
harfbuzz=1.8.8=hffaf4a1_0
hdf5=1.10.2=hba1933b_1
hydra-core=1.1.1=pyhd8ed1ab_0
icu=58.2=he6710b0_3
idna=2.10=pyhd3eb1b0_0
imagecodecs=2021.1.11=py37h581e88b_1
imageio=2.9.0=py_0
importlib_resources=5.7.1=pyhd8ed1ab_1
intel-openmp=2019.4=243
ipython=7.21.0=py37hb070fc8_0
ipython_genutils=0.2.0=pyhd3eb1b0_1
jasper=2.0.14=h07fcdf6_1
jedi=0.17.0=py37_0
joblib=1.0.1=pyhd3eb1b0_0
jpeg=9b=h024ee3a_2
jxrlib=1.1=h7b6447c_2
kiwisolver=1.3.1=py37h2531618_0
lcms2=2.11=h396b838_0
ld_impl_linux-64=2.35.1=h2b15070_1
lerc=2.2.1=h2531618_0
leveldb=1.20=hf484d3e_1
libaec=1.0.4=he6710b0_1
libboost=1.67.0=h46d08c1_4
libdeflate=1.7=h27cfd23_5
libedit=3.1.20191231=h14c3975_1
libffi=3.3=he6710b0_2
libgcc-ng=9.3.0=h9cbead1_17
libgfortran-ng=7.5.0=hae1eefd_17
libgfortran4=7.5.0=hae1eefd_17
libglu=9.0.0=hf484d3e_1
libgomp=9.3.0=h9cbead1_17
libmklml=2019.0.5=0
libopencv=3.4.2=hb342d67_1
libopus=1.3.1=h7b6447c_0
libpng=1.6.37=hbc83047_0
libprotobuf=3.15.4=h780b84a_0
libstdcxx-ng=9.3.0=h2ae2ef3_17
libtiff=4.1.0=h2733197_1
libuuid=1.0.3=h1bed415_2
libuv=1.40.0=h7b6447c_0
libvpx=1.7.0=h439df22_0
libwebp=1.0.1=h8e7db2f_0
libxcb=1.14=h7b6447c_0
libxml2=2.9.10=hb55368b_3
libzopfli=1.0.3=he6710b0_0
lmdb=0.9.28=h2531618_0
lz4-c=1.9.3=h2531618_0
matplotlib=3.3.2=h06a4308_0
matplotlib-base=3.3.2=py37h817c723_0
mkl=2020.2=256
mkl-service=2.3.0=py37he8ac12f_0
mkl_fft=1.2.0=py37h23d657b_0
mkl_random=1.1.1=py37h0573a6f_0
ncurses=6.2=he6710b0_1
networkx=2.5=py_0
ninja=1.10.2=py37hff7bd54_0
numpy=1.19.2=py37h54aff64_0
numpy-base=1.19.2=py37hfa32c7d_0
olefile=0.46=py_0
omegaconf=2.1.1=py37h89c1867_1
openjpeg=2.3.0=h05c96fa_1
openssl=1.1.1m=h7f8727e_0
pandas=1.2.3=py37ha9443f7_0
parso=0.8.1=pyhd3eb1b0_0
pcre=8.44=he6710b0_0
pexpect=4.8.0=pyhd3eb1b0_3
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=8.1.0=py37he98fc37_0
pip=20.3.3=py37h06a4308_0
pixman=0.40.0=h7b6447c_0
prompt-toolkit=3.0.8=py_0
protobuf=3.15.4=py37hcd2ae1e_0
psutil=5.8.0=py37h27cfd23_1
ptyprocess=0.7.0=pyhd3eb1b0_2
py-boost=1.67.0=py37h04863e7_4
py-opencv=3.4.2=py37hb342d67_1
pycparser=2.20=py_2
pygments=2.8.0=pyhd3eb1b0_0
pyopenssl=20.0.1=pyhd3eb1b0_1
pyparsing=2.4.7=pyhd3eb1b0_0
pyqt=5.9.2=py37h05f1152_2
pysocks=1.7.1=py37_1
python=3.7.9=h7579374_0
python-dateutil=2.8.1=pyhd3eb1b0_0
python-gflags=3.1.2=py37h06a4308_0
python-leveldb=0.201=py37h2531618_0
python_abi=3.7=1_cp37m
pytorch=1.7.1=py3.7_cuda11.0.221_cudnn8.0.5_0
pytz=2021.1=pyhd3eb1b0_0
pywavelets=1.1.1=py37h7b6447c_2
pyyaml=5.4.1=py37h27cfd23_1
qt=5.9.7=h5867ecd_1
readline=8.1=h27cfd23_0
requests=2.25.1=pyhd3eb1b0_0
scikit-image=0.17.2=py37hdf5156a_0
scikit-learn=0.24.1=py37ha9443f7_0
scipy=1.1.0=py37h7c811a0_2
seaborn=0.11.1=pyhd3eb1b0_0
selenium=3.141.0=py37h14c3975_1001
setuptools=52.0.0=py37h06a4308_0
sip=4.19.8=py37hf484d3e_0
six=1.15.0=pyhd3eb1b0_0
sk-video=1.1.10=pyh24bf2e0_4
snappy=1.1.8=he6710b0_0
sqlite=3.33.0=h62c20be_0
tensorboardx=2.1=py_0
threadpoolctl=2.1.0=pyh5ca1d4c_0
tifffile=2021.1.14=pyhd3eb1b0_1
tk=8.6.10=hbc83047_0
toolz=0.11.1=pyhd3eb1b0_0
torchaudio=0.7.2=py37
torchvision=0.8.2=py37_cu110
tornado=6.1=py37h27cfd23_0
tqdm=4.59.0=pyhd3eb1b0_1
traitlets=5.0.5=pyhd3eb1b0_0
typing-extensions=3.7.4.3=hd3eb1b0_0
typing_extensions=3.7.4.3=pyh06a4308_0
urllib3=1.26.3=pyhd3eb1b0_0
wcwidth=0.2.5=py_0
wheel=0.36.2=pyhd3eb1b0_0
xz=5.2.5=h7b6447c_0
yacs=0.1.6=pyhd3eb1b0_1
yaml=0.2.5=h7b6447c_0
zfp=0.5.5=h2531618_4
zipp=3.8.0=pyhd8ed1ab_0
zlib=1.2.11=h7b6447c_3
zstd=1.4.5=h9ceee32_0
